Trait vm::common::build_cfg::ControlFlowGraph

pub trait ControlFlowGraph: Sized {
    fn new(code: &[Bytecode]) -> Result<Self, VMStaticViolation>;

    fn reachable_from(&self, block_id: BlockId) -> Vec<&BasicBlock>;

    fn block_id_of_offset(&self, code_offset: CodeOffset) -> Option<BlockId>;

    fn block_of_id(&self, block_id: BlockId) -> Option<&BasicBlock>;

    fn num_blocks(&self) -> u16;
}

A trait that specifies the basic requirements for a CFG

Required methods

fn new(code: &[Bytecode]) -> Result<Self, VMStaticViolation>

Given a vector of bytecodes, constructs the control flow graph for it. Return a VMStaticViolation if we were unable to construct a control flow graph, or if we encounter an invalid jump instruction.

fn reachable_from(&self, block_id: BlockId) -> Vec<&BasicBlock>

Given a block ID, return the reachable blocks from that block including the block itself.

fn block_id_of_offset(&self, code_offset: CodeOffset) -> Option<BlockId>

Given an offset into the bytecode return the basic block ID that contains that offset

fn block_of_id(&self, block_id: BlockId) -> Option<&BasicBlock>

Given a block ID, return the corresponding basic block. Return None if the block ID is invalid.

fn num_blocks(&self) -> u16

Return the number of blocks (vertices) in the control flow graph

Loading content...