Module vm::errors

Structs

Location

Constants

EACCOUNT_DOES_NOT_EXIST

EACCOUNT_FROZEN
Error codes that can be emitted by the prologue. These have special significance to the VM when they are raised during the prologue. However, they can also be raised by user code during execution of a transaction script. They have no significance to the VM in that case.

EASSERT_ERROR

EBAD_ACCOUNT_AUTHENTICATION_KEY

EBAD_SIGNATURE

ECANT_PAY_GAS_DEPOSIT

EINSUFFICIENT_BALANCE
Generic error codes. These codes don't have any special meaning for the VM, but they are useful conventions for debugging

EINSUFFICIENT_PRIVILEGES

ESEQUENCE_NUMBER_TOO_NEW

ESEQUENCE_NUMBER_TOO_OLD

ETRANSACTION_EXPIRED

Functions

append_err_info

bounds_error

bytecode_offset_err

convert_prologue_runtime_error

err_at_offset

verification_error

vm_error

vm_result_to_transaction_status

vm_status_of_result
Conversion functions from internal VM statuses into external VM statuses

Type Definitions

BinaryLoaderResult

VMResult