Module vm::file_format

Structs

AddressPoolIndex

ByteArrayPoolIndex

CodeUnit

CompiledModule

CompiledScript

FieldDefinition

FieldDefinitionIndex

FunctionDefinition

FunctionDefinitionIndex

FunctionHandle

FunctionHandleIndex

FunctionSignature
A function signature serializes the return type as its first element and argument types as its second element.

FunctionSignatureIndex

LocalsSignature
Represents signatures for local variables within a function.

LocalsSignatureIndex

ModuleHandle

ModuleHandleIndex

StringPoolIndex

StructDefinition

StructDefinitionIndex

StructHandle

StructHandleIndex

TypeSignature
Signature definitions. A signature can be for a type (field, local) or for a function - return type: (arguments). They both go into the signature table so there is a marker that tags the signature. Signature usually don't carry a size and you have to read them to get to the end.

TypeSignatureIndex

Enums

Bytecode

SignatureToken

Constants

SELF_MODULE_NAME

Type Definitions

AddressPool

ByteArrayPool

CodeOffset

FunctionSignaturePool
The pool of FunctionSignature instances.

LocalIndex

LocalsSignaturePool
The pool of LocalsSignature instances.

MemberCount

StringPool

TableIndex

TypeSignaturePool
The pool of TypeSignature instances.