Struct vm::file_format::AbilitySet

pub struct AbilitySet(_);

A set of Abilitys

Implementations

impl AbilitySet

pub const EMPTY: Self

The empty ability set

pub const PRIMITIVES: AbilitySet

Abilities for Bool, U8, U64, U128, and Address

pub const REFERENCES: AbilitySet

Abilities for Reference and MutableReference

pub const SIGNER: AbilitySet

Abilities for Signer

pub const VECTOR: AbilitySet

Abilities for Vector, note they are predicated on the type argument

pub const ALL: Self

Ability set containing all abilities

pub fn has_ability(self, ability: Ability) -> bool

pub fn has_copy(self) -> bool

pub fn has_drop(self) -> bool

pub fn has_store(self) -> bool

pub fn has_key(self) -> bool

pub fn remove(self, ability: Ability) -> Self

pub fn intersect(self, other: Self) -> Self

pub fn union(self, other: Self) -> Self

pub fn is_subset(self, other: Self) -> bool

pub fn polymorphic_abilities(
declared_abilities: Self,
type_argument_abilities: impl IntoIterator<Item = Self>
) -> Self

For a polymorphic type, its actual abilities correspond to its declared abilities but predicated on its type arguments having that ability. For Key, instead of needing the same ability, the type arguments need Store

pub fn from_u8(byte: u8) -> Option<Self>

pub fn into_u8(self) -> u8