Struct vm::file_format::CompiledModule

pub struct CompiledModule(_);

A CompiledModule defines the structure of a module which is the unit of published code.

A CompiledModule contains a definition of types (with their fields) and functions. It is a unit of code that can be used by transactions or other modules.

A module is published as a single entry and it is retrieved as a single blob.

Methods

impl CompiledModule

pub fn deserialize(binary: &[u8]) -> BinaryLoaderResult<Self>

Deserialize a &u8 slice into a CompiledModule instance.

impl CompiledModule

pub fn as_inner(&self) -> &CompiledModuleMut

Returns a reference to the inner CompiledModuleMut.

pub fn into_inner(self) -> CompiledModuleMut

Converts this instance into the inner CompiledModuleMut. Converting back to a CompiledModule would require it to be verified again.

pub fn kind_count(&self, kind: IndexKind) -> usize

Returns the number of items of a specific IndexKind.

pub fn module_id_for_handle(&self, module_handle: &ModuleHandle) -> ModuleId

Returns the code key of module_handle

pub fn self_id(&self) -> ModuleId

Returns the code key of self

pub fn into_script(self, conv_info: ScriptConversionInfo) -> CompiledScript

Deprecated:
This function is deprecated and will be removed soon. Please do not introduce new dependencies.

This function should only be called on an instance of CompiledModule obtained by invoking into_module on some instance of CompiledScript. This function is the inverse of into_module, i.e., script.into_module().into_script() == script.

impl CompiledModule

pub fn valid_strategy(size: usize) -> impl Strategy<Value = Self>

Convenience wrapper around CompiledModuleStrategyGen that generates valid modules with the given size.

impl CompiledModule

pub fn serialize(&self, binary: &mut Vec<u8>) -> Result<()>

Serializes a CompiledModule into a binary. The mutable Vec<u8> will contain the binary blob on return.