Struct vm::file_format::CompiledModuleMut

pub struct CompiledModuleMut {
    pub module_handles: Vec<ModuleHandle>,
    pub struct_handles: Vec<StructHandle>,
    pub function_handles: Vec<FunctionHandle>,
    pub type_signatures: TypeSignaturePool,
    pub function_signatures: FunctionSignaturePool,
    pub locals_signatures: LocalsSignaturePool,
    pub identifiers: IdentifierPool,
    pub user_strings: UserStringPool,
    pub byte_array_pool: ByteArrayPool,
    pub address_pool: AddressPool,
    pub struct_defs: Vec<StructDefinition>,
    pub field_defs: Vec<FieldDefinition>,
    pub function_defs: Vec<FunctionDefinition>,
}

A mutable version of CompiledModule. Converting to a CompiledModule requires this to pass the bounds checker.

Fields
module_handles: Vec<ModuleHandle>

Handles to external modules and self at position 0.

struct_handles: Vec<StructHandle>

Handles to external and internal types.

function_handles: Vec<FunctionHandle>

Handles to external and internal functions.

type_signatures: TypeSignaturePool

Type pool. A definition for all types used in the module.

function_signatures: FunctionSignaturePool

Function signature pool. Represents all function signatures defined or used in the module.

locals_signatures: LocalsSignaturePool

Locals signature pool. The signature for all locals of the functions defined in the module.

identifiers: IdentifierPool

All identifiers used in this module.

user_strings: UserStringPool

User strings. All literals used in this module.

byte_array_pool: ByteArrayPool

ByteArray pool. The byte array literals used in the module.

address_pool: AddressPool

Address pool. The address literals used in the module. Those include literals for code references (ModuleHandle).

struct_defs: Vec<StructDefinition>

Types defined in this module.

field_defs: Vec<FieldDefinition>

Fields defined on types in this module.

function_defs: Vec<FunctionDefinition>

Function defined in this module.

Methods

impl CompiledModuleMut

pub fn deserialize_no_check_bounds(binary: &[u8]) -> BinaryLoaderResult<Self>

impl CompiledModuleMut

pub fn kind_count(&self, kind: IndexKind) -> usize

Returns the count of a specific IndexKind

pub fn freeze(self) -> Result<CompiledModule, Vec<VMStatus>>

Converts this instance into CompiledModule after verifying it for basic internal consistency. This includes bounds checks but no others.

impl CompiledModuleMut

pub fn serialize(&self, binary: &mut Vec<u8>) -> Result<()>

Serializes this into a binary format.

This is intended mainly for test code. Production code will typically use [CompiledModule::serialize].