Enum vm::gas_schedule::GasCostTable

pub enum GasCostTable {
    Low,
    Mid,
    High,
    EmitEvent,
    GetSequenceNumber,
    GetPublicKey,
    GetGasUnitPrice,
    GetMaxGasUnits,
    GetGasRemaining,
    GetSenderAddress,
    Unpack,
    Exists,
    BorrowGlobal,
    ReleaseRef,
    MoveFrom,
    MoveToSender,
    CreateAccount,
}

For now, we separate operations into three different abstract tiers, each tier representing the differing cost of performing a single ('atomic') operation. We then have a number of others for the various builtin functions.

Variants
LowMidHighEmitEventGetSequenceNumberGetPublicKeyGetGasUnitPriceGetMaxGasUnitsGetGasRemainingGetSenderAddressUnpackExistsBorrowGlobalReleaseRefMoveFromMoveToSenderCreateAccount
Methods

impl GasCostTable

We now provide denotations for these abstract gas tiers. Note that these denotations are filler (read: bogus) for now. The constants will be filled in with something more meaningful later on, or provided by a separate contract that we read in.

pub fn instr_cost(&self) -> GasUnits

pub fn mem_cost(&self) -> GasUnits

pub fn stack_cost(&self) -> GasUnits