Struct vm::PriceList

pub struct PriceList {
    pub on_chain_message_base: i64,
    pub on_chain_message_per_byte: i64,
    pub on_chain_return_value_per_byte: i64,
    pub send_base: i64,
    pub send_transfer_funds: i64,
    pub send_invoke_method: i64,
    pub ipld_get_base: i64,
    pub ipld_get_per_byte: i64,
    pub ipld_put_base: i64,
    pub ipld_put_per_byte: i64,
    pub create_actor_base: i64,
    pub create_actor_extra: i64,
    pub delete_actor: i64,
    pub hashing_base: i64,
    pub hashing_per_byte: i64,
    pub compute_unsealed_sector_cid_base: i64,
    pub verify_seal_base: i64,
    pub verify_post_base: i64,
    pub verify_consensus_fault: i64,
}

Provides prices for operations in the VM

Fields
on_chain_message_base: i64

Gas cost charged to the originator of an on-chain message (regardless of whether it succeeds or fails in application) is given by: OnChainMessageBase + len(serialized message)*OnChainMessagePerByte Together, these account for the cost of message propagation and validation, up to but excluding any actual processing by the VM. This is the cost a block producer burns when including an invalid message.

on_chain_message_per_byte: i64on_chain_return_value_per_byte: i64

Gas cost charged to the originator of a non-nil return value produced by an on-chain message is given by: len(return value)*OnChainReturnValuePerByte

send_base: i64

Gas cost for any message send execution(including the top-level one initiated by an on-chain message). This accounts for the cost of loading sender and receiver actors and (for top-level messages) incrementing the sender's sequence number. Load and store of actor sub-state is charged separately.

send_transfer_funds: i64

Gas cost charged, in addition to SendBase, if a message send is accompanied by any nonzero currency amount. Accounts for writing receiver's new balance (the sender's state is already accounted for).

send_invoke_method: i64

Gas cost charged, in addition to SendBase, if a message invokes a method on the receiver. Accounts for the cost of loading receiver code and method dispatch.

ipld_get_base: i64

Gas cost (Base + len*PerByte) for any Get operation to the IPLD store in the runtime VM context.

ipld_get_per_byte: i64ipld_put_base: i64

Gas cost (Base + len*PerByte) for any Put operation to the IPLD store in the runtime VM context. Note: these costs should be significantly higher than the costs for Get operations, since they reflect not only serialization/deserialization but also persistent storage of chain data.

ipld_put_per_byte: i64create_actor_base: i64

Gas cost for creating a new actor (via InitActor's Exec method). Note: this costs assume that the extra will be partially or totally refunded while the base is covering for the put.

create_actor_extra: i64delete_actor: i64

Gas cost for deleting an actor. Note: this partially refunds the create cost to incentivise the deletion of the actors.

hashing_base: i64hashing_per_byte: i64compute_unsealed_sector_cid_base: i64verify_seal_base: i64verify_post_base: i64verify_consensus_fault: i64
Methods

impl PriceList

pub fn on_chain_message(&self, msg_size: i64) -> i64

Returns the gas required for storing a message of a given size in the chain.

pub fn on_chain_return_value(&self, data_size: usize) -> i64

Returns the gas required for storing the response of a message in the chain.

pub fn on_method_invocation(
&self,
value: &TokenAmount,
method_num: MethodNum
) -> i64

Returns the gas required when invoking a method.

pub fn on_ipld_get(&self, data_size: usize) -> i64

Returns the gas required for storing an object

pub fn on_ipld_put(&self, data_size: usize) -> i64

Returns the gas required for storing an object

pub fn on_create_actor(&self) -> i64

Returns the gas required for creating an actor

pub fn on_delete_actor(&self) -> i64

Returns the gas required for deleting an actor

pub fn on_verify_signature(
&self,
sig_type: SignatureType,
plain_text_size: usize
) -> i64

Returns gas required for signature verification

pub fn on_hashing(&self, data_size: usize) -> i64

Returns gas required for hashing data

pub fn on_compute_unsealed_sector_cid(
&self,
_proof: RegisteredProof,
_pieces: &[PieceInfo]
) -> i64

Returns gas required for computing unsealed sector Cid

pub fn on_verify_seal(&self, _info: &SealVerifyInfo) -> i64

Returns gas required for seal verification

pub fn on_verify_post(&self, _info: &WindowPoStVerifyInfo) -> i64

Returns gas required for PoSt verification

pub fn on_verify_consensus_fault(&self) -> i64

Returns gas required for verifying consensus fault