Struct vm::tests::FakeExt

pub struct FakeExt {

    pub initial_store: HashMap<H256, H256>,
    pub store: HashMap<H256, H256>,
    pub suicides: HashSet<Address>,
    pub calls: HashSet<FakeCall>,
    pub sstore_clears: i128,
    pub depth: usize,
    pub blockhashes: HashMap<U256, H256>,
    pub codes: HashMap<Address, Arc<Bytes>>,
    pub logs: Vec<FakeLogEntry>,
    pub info: EnvInfo,
    pub schedule: Schedule,
    pub balances: HashMap<Address, U256>,
    pub tracing: bool,
    pub is_static: bool,
    pub access_list: AccessList,
    /* private fields */

}

Fake externalities test structure.

Can’t do recursive calls.

Fields
initial_store: HashMap<H256, H256>store: HashMap<H256, H256>suicides: HashSet<Address>calls: HashSet<FakeCall>sstore_clears: i128depth: usizeblockhashes: HashMap<U256, H256>codes: HashMap<Address, Arc<Bytes>>logs: Vec<FakeLogEntry>info: EnvInfoschedule: Schedulebalances: HashMap<Address, U256>tracing: boolis_static: boolaccess_list: AccessList
Implementations

impl FakeExt

pub fn new() -> Self

New fake externalities

pub fn new_byzantium() -> Self

New fake externalities with byzantium schedule rules

pub fn new_constantinople() -> Self

New fake externalities with constantinople schedule rules

pub fn new_istanbul() -> Self

New fake externalities with Istanbul schedule rules

pub fn new_berlin(from: Address, to: Address, builtins: &[Address]) -> Self

New fake externalities with Berlin schedule rules

pub fn new_london(from: Address, to: Address, builtins: &[Address]) -> Self

New fake externalities with London schedule rules

pub fn with_wasm(self) -> Self

Alter fake externalities to allow wasm

pub fn with_chain_id(self, chain_id: u64) -> Self

Set chain ID

pub fn set_initial_storage(&mut self, key: H256, value: H256)

pub fn prefill(&mut self, data: &[u64])

Fill the storage before the transaction with data, i. e. set both original and current values, beginning from address 0.