Module vm::views

An alternate representation of the file format built on top of the existing format.

Some general notes:

These views are not meant to be set in stone. Feel free to change the views exposed as the format and our understanding evolves.

The typical use for these views would be to materialize all the lazily evaluated data immediately -- the views are a convenience to make that simpler. They've been written as lazy iterators to aid understanding of the file format and to make it easy to generate views.

Structs

FieldDefinitionView

FunctionDefinitionView

FunctionHandleView

FunctionSignatureView

LocalsSignatureView

ModuleHandleView

ModuleView
Represents a lazily evaluated abstraction over a module.

SignatureTokenView

StructDefinitionView

StructHandleView

TypeSignatureView

Traits

ViewInternals
This is used to expose some view internals to checks and other areas. This might be exposed to external code in the future.