Crate vm

Re-exports

pub use file_format::CompiledModule;

Modules

access
Defines accessors for compiled modules.

check_bounds

constant

deserializer

errors

file_format
Binary format for transactions and modules.

file_format_common
Constants for the binary format.

internals
Types meant for use by other parts of this crate, and by other crates that are designed to work with the internals of these data structures.

proptest_types
Utilities for property-based testing.

serializer
Serialization of transactions and modules.

views
An alternate representation of the file format built on top of the existing format.

Enums

IndexKind
Represents a kind of index -- useful for error messages.

SignatureTokenKind
Represents the kind of a signature token.