Crate vm

Structs

ActorID
Identifier for Actors, includes builtin and initialized actors

ActorState
State of all actor implementations

InvocInput
Input variables for actor method invocation.

MethodNum
Method number indicator for calling actor methods

PaddedPieceSize
Size of a piece in bytes with padding

PieceInfo

PoStVerifyInfo

SealVerifyInfo

Serialized
Serialized bytes to be used as parameters into actor methods

TokenAmount
Wrapper around a big int variable to handle token specific functionality

UnpaddedPieceSize
Size of a piece in bytes

Enums

CodeID
CodeID is the reference to the code which is attached to the Actor state. There are builtin IDs and the option for custom code with a Cid

ExitCode
ExitCode defines the exit code from the VM execution

RegisteredProof

Constants

METHOD_CONSTRUCTOR
Base actor constructor method

METHOD_SEND
Base actor send method

Traits

StateTree
Interface to allow for the retreival and modification of Actors and their state