Crate vm

Structs

InvocInput
Input variables for actor method invocation.

InvocOutput
Output variables for actor method invocation.

MethodNum
Method number indicator for calling actor methods

Serialized
Serialized bytes to be used as parameters into actor methods

TokenAmount
Wrapper around a big int variable to handle token specific functionality

Enums

ExitCode
ExitCode defines the exit code from the VM execution

SysCode
Defines the system error codes defined by the protocol

UserCode
defines user specific error codes from VM execution

Constants

METHOD_CONSTRUCTOR
Base actor constructor method

METHOD_CRON
Base actor cron method

METHOD_PLACEHOLDER
Placeholder for non base methods for actors

METHOD_SEND
Base actor send method