Crate vm

Virtual machines support library

Re-exports

pub use access_list::AccessList;

pub use schedule::CleanDustMode;

pub use schedule::Schedule;

pub use schedule::WasmCosts;

Modules

access_list

schedule

Cost schedule and other parameterisations for the EVM.

tests

Structs

ActionParams

Action (call/create) input params. Everything else should be specified in Externalities.

EnvInfo

Information concerning the execution environment for a message-call/contract-creation.

ReturnData

Return data buffer. Holds memory from a previous call and a slice into that memory.

Enums

ActionValue

Transaction value

CallType

The type of the call-like instruction.

ContractCreateResult

Result of externalities create function.

CreateContractAddress

Specifies how an address is calculated for a new contract.

Error

VM errors.

GasLeft

Gas Left: either it is a known value, or it needs to be computed by processing a return instruction.

MessageCallResult

Result of externalities call function.

ParamsType

Type of the way parameters encoded

TrapError

TrapKind

Traits

Exec

Virtual Machine interface

Ext

Externalities interface for EVMs

ResumeCall

Resume call interface

ResumeCreate

Resume create interface

Type Definitions

ExecTrapError

ExecTrapResult

LastHashes

Simple vector of hashes, should be at most 256 items large, can be smaller if being used for a block whose number is less than 257.

Result

TrapResult