Crate vm

Structs

ActorError
The error type that gets returned by actor method calls.

ActorState
State of all actor implementations

GasTracker

InvocInput
Input variables for actor method invocation.

OnChainSealVerifyInfo
OnChainSealVerifyInfo is the structure of information that must be sent with a message to commit a sector. Most of this information is not needed in the state tree but will be verified in sm.CommitSector. See SealCommitment for data stored on the state tree for each sector.

OnChainWindowPoStVerifyInfo
Information submitted by a miner to provide a Window PoSt.

PaddedPieceSize
Size of a piece in bytes with padding

PieceInfo

PoStProof

PriceList
Provides prices for operations in the VM

SealVerifyInfo
Information needed to verify a seal proof.

SectorID
Sector ID which contains the sector number and the actor ID for the miner.

SectorInfo
Information about a sector necessary for PoSt verification

Serialized
Serialized bytes to be used as parameters into actor methods

UnpaddedPieceSize
Size of a piece in bytes

WindowPoStVerifyInfo
Information needed to verify a Window PoSt submitted directly to a miner actor.

WinningPoStVerifyInfo
Information needed to verify a Winning PoSt attached to a block header. Note: this is not used within the state machine, but by the consensus/election mechanisms.

Enums

CodeID
CodeID is the reference to the code which is attached to the Actor state. There are builtin IDs and the option for custom code with a Cid

ExitCode
ExitCode defines the exit code from the VM execution

RegisteredProof
This ordering, defines mappings to UInt in a way which MUST never change.

SectorSize
SectorSize indicates one of a set of possible sizes in the network.

Constants

METHOD_CONSTRUCTOR
Base actor constructor method

METHOD_SEND
Base actor send method

Functions

price_list_by_epoch
Returns gas price list by Epoch for gas consumption

zero_piece_commitment

Type Definitions

ActorID
Identifier for Actors, includes builtin and initialized actors

DealID
Deal identifier used in market and miner actors

InteractiveSealRandomness

MethodNum
Method number indicator for calling actor methods

PoStRandomness

Randomness
String of random bytes

SealRandomness

SectorNumber

StoragePower
Unit of storage power (measured in bytes)

TokenAmount